+++ apps/qqtx.js
@@ -20,15 +20,20 @@ export class qqtx extends plugin {
     }
     async qqtx(e) {
         if (CONFIG_YAML.qqtx == false) {
-            logger.error('举牌已关闭');
+            logger.error('QQ头像已关闭');
             return false
         }
+        try {
         let user_id = e.at || e.user_id
         user_id = Number(user_id) || String(user_id)
         let data = await fs.readFileSync(`${Plugin_Path}/config/AllAPI.json`)
         const API = JSON.parse(data)
-        let api = API.api10.url + `?qq=${user_id}`
+        let api = API.api10.url + `?b=qq&nk=${user_id}&s=640`
          e.reply([segment.image(`${api}`)])
+        } catch (err) {
+            logger.error(`获取QQ头像时出错: ${err.message}`);
+            e.reply('获取QQ头像失败，请稍后再试。');
+        }
          return true
      }
  }